Revision: Make phone/email parameters optional
The current AddressBook required users to fill in both the phone number
and the email address. This meant that contacts who do not have both a
phone numbe and email address could not be added into the AddressBook.

This commit addresses this issue by making the phone and email fields
optional.